requirements.txt
asgiref==3.11.0
colorama==0.4.6
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
sqlparse==0.5.5
tzdata==2025.3
